Add locale changing feature
Create SetLocale middleware and assign this to all routes.
When a request is sent to the server, SetLocale middleware reads the
locale value from these sources: session -> cookie -> configured locale.

Create a route for handling locale changing request.
When received a locale changing request, save the locale value in the
session.